+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * #4232: Allow configuring standard authentication service in HA mode
 * #4455: Qpid Dispatch Router upgraded to 1.12.0
 * #4504: Bump gqlgen from v0.10.2 => v0.11.3
+* #4546: Refactor GraphQL purgeAddress to permit multiple target addresses
 
 ## 0.31.2
 * #4098: Added alert for high broker address memory usage